[refactor] 패키징 변경
util/binding 으로 폴더링
