+++ app/src/main/java/com/example/healthhub/util/binding/BindingActivity.kt
@@ -1,4 +1,4 @@
-package com.example.healthhub.util
+package com.example.healthhub.util.binding
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
